old compiler and os fix

+++ ssl/ech/ech_internal.c
@@ -805,7 +805,7 @@ void ossl_ech_status_print(BIO *out, SSL_CONNECTION *s, int selector)
                 continue;
             }
             BIO_printf(out, "ECH entry: %d public_name: %s age: %lld%s\n",
-                i, pn, (long long)secs, has_priv ? " (has private key)" : "");
+                i, pn, (int64_t)secs, has_priv ? " (has private key)" : "");
             BIO_printf(out, "\t%s\n", ec);
             OPENSSL_free(pn);
             OPENSSL_free(ec);
